Test cases added
If stackIdx>input.size()-1 from the get-go,  we cannot recurse further. Thus, we return negative(not mathematically la) if the isSolution() is not true.
BUT, we memoize that, we may have FALSE negative results for some of the nodes in the recursion tree (for that particular path).

I don't know to counter this. The code i added can handle only some of the test cases.